+++ plugins/cocart/includes/classes/rest-api/class-cocart-rest-response.php
@@ -30,38 +30,7 @@ class CoCart_Response {
 	 * @ignore Function ignored when parsed into Code Reference.
 	 */
 	public function __construct() {
-		// Expose custom headers.
-		add_action( 'rest_pre_serve_request', array( $this, 'expose_custom_headers' ), 11, 4 );
 	} // END __construct()
-
-	/**
-	 * Expose CoCart Headers.
-	 *
-	 * @access public
-	 *
-	 * @since 3.1.0 Introduced.
-	 * @since 3.3.0 Added new custom headers without the prefix `X-`
-	 * @since 4.0.0 Removed old custom headers with the prefix `X-`
-	 *
-	 * @param bool             $served  Whether the request has already been served. Default false.
-	 * @param WP_HTTP_Response $result  Result to send to the client. Usually a WP_REST_Response.
-	 * @param WP_REST_Request  $request Request used to generate the response.
-	 * @param WP_REST_Server   $server  Server instance.
-	 *
-	 * @return bool Whether the request has already been served.
-	 */
-	public function expose_custom_headers( $served, $result, $request, $server ) {
-		if ( strpos( $request->get_route(), 'cocart/' ) !== false ) {
-			header( 'Access-Control-Expose-Headers: CoCart-Timestamp' );
-			header( 'Access-Control-Expose-Headers: CoCart-Version' );
-			header( 'Access-Control-Expose-Headers: CoCart-API-Cart-Key' );
-			header( 'Access-Control-Expose-Headers: CoCart-API-Customer' );
-			header( 'Access-Control-Expose-Headers: CoCart-API-Cart-Expiring' );
-			header( 'Access-Control-Expose-Headers: CoCart-API-Cart-Expiration' );
-		}
-
-		return $served;
-	} // END expose_custom_headers()
 
 	/**
 	 * Returns either the default response of the API requested or a filtered response.
