+++ plugins/cocart/includes/classes/rest-api/controllers/v2/cart/class-cocart-rest-cart-v2-controller.php
@@ -12,10 +12,9 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+//use CoCart\Abstracts;
+use CoCart\Schemas\v2\CartSchema as Schema;
 use CoCart\Session\Handler;
-use CoCart\RestApi\CartCache;
-
-use \Automattic\WooCommerce\Checkout\Helpers\ReserveStock;
 
 /**
  * Main cart controller that gets the requested cart in session
@@ -27,9 +26,9 @@ use \Automattic\WooCommerce\Checkout\Helpers\ReserveStock;
  *
  * @since 3.0.0 Introduced.
  *
- * @see CoCart_API_Controller
+ * @see CoCart_Cart_REST_Controller
  */
-class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
+class CoCart_REST_Cart_v2_Controller extends CoCart_Cart_REST_Controller {
 
 	/**
 	 * Endpoint namespace.
@@ -46,49 +45,32 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	protected $rest_base = 'cart';
 
 	/**
-	 * Register the routes for cart.
-	 *
-	 * @access public
-	 *
-	 * @ignore Function ignored when parsed into Code Reference.
+	 * Constructor.
 	 */
-	public function register_routes() {
-		// Get Cart - cocart/v2/cart (GET).
-		register_rest_route(
-			$this->namespace,
-			'/' . $this->rest_base,
-			array(
-				array(
-					'methods'             => WP_REST_Server::READABLE,
-					'callback'            => array( $this, 'get_cart' ),
-					'permission_callback' => '__return_true',
-					'args'                => $this->get_collection_params(),
-				),
-				'schema' => array( $this, 'get_public_item_schema' ),
-			)
-		);
-	} // register_routes()
+	public function __construct() {
+		$this->schema = new Schema;
+	}
 
 	/**
-	 * Gets the cart instance so we only call it once in the API.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
+	 * Get method arguments for this REST route.
 	 *
 	 * @access public
 	 *
-	 * @since 3.0.0 Introduced.
+	 * @since 4.0.0 Introduced.
 	 *
-	 * @return WC_Cart
+	 * @return array An array of endpoints.
 	 */
-	public function get_cart_instance() {
-		$cart = WC()->cart;
-
-		if ( ! $cart || ! $cart instanceof \WC_Cart ) {
-			throw new CoCart_Data_Exception( 'cocart_cart_error', __( 'Unable to retrieve cart.', 'cart-rest-api-for-woocommerce' ), 404 );
-		}
-
-		return $cart;
-	} // END get_cart_instance()
+	public function get_args() {
+		return array(
+			array(
+				'methods'             => WP_REST_Server::READABLE,
+				'callback'            => array( $this, 'get_cart' ),
+				'permission_callback' => '__return_true',
+				'args'                => $this->get_collection_params(),
+			),
+			'schema' => array( $this, 'get_public_item_schema' ),
+		);
+	} // END get_args()
 
 	/**
 	 * Return a cart item from the cart.
@@ -120,49 +102,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	} // EMD get_cart_item()
 
 	/**
-	 * Returns all cart items.
-	 *
-	 * @access public
-	 *
-	 * @param callable $callback Optional callback to apply to the array filter.
-	 *
-	 * @return array $items Returns all cart items.
-	 */
-	public function get_cart_items( $callback = null ) {
-		return $callback ? array_filter( $this->get_cart_instance()->get_cart(), $callback ) : array_filter( $this->get_cart_instance()->get_cart() );
-	} // END get_cart_items()
-
-	/**
-	 * Returns true if the cart is completely empty.
-	 *
-	 * @access public
-	 *
-	 * @since 3.1.0 Introduced.
-	 *
-	 * @return bool True if the cart is completely empty.
-	 */
-	public function is_completely_empty() {
-		if ( $this->get_cart_instance()->get_cart_contents_count() <= 0 && $this->get_removed_cart_contents_count() <= 0 ) {
-			return true;
-		}
-
-		return false;
-	} // END is_completely_empty()
-
-	/**
-	 * Get number of removed items in the cart.
-	 *
-	 * @access public
-	 *
-	 * @since 3.1.0 Introduced.
-	 *
-	 * @return int Number of removed items in the cart.
-	 */
-	public function get_removed_cart_contents_count() {
-		return array_sum( wp_list_pluck( $this->get_cart_instance()->get_removed_cart_contents(), 'quantity' ) );
-	} // END get_removed_cart_contents_count()
-
-	/**
 	 * Get cart.
 	 *
 	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
@@ -172,12 +111,12 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	 * @since   3.0.0 Introduced.
 	 * @version 4.0.0
 	 *
-	 * @param WP_REST_Request $request       Full details about the request.
-	 * @param string          $cart_item_key Originally the cart item key.
+	 * @param \WP_REST_Request $request       Request object.
+	 * @param string           $cart_item_key Originally the cart item key.
 	 *
 	 * @return WP_REST_Response $response The response data.
 	 */
-	public function get_cart( $request = array(), $cart_item_key = null ) {
+	public function get_cart( \WP_REST_Request $request, $cart_item_key = null ) {
 		try {
 			// Checks that we are using CoCart session handler. If not detected, throw error response.
 			if ( ! WC()->session instanceof Handler ) {
@@ -219,9 +158,9 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 
 			$cart_contents = $this->return_cart_contents( $request, $cart_contents );
 
-			return CoCart_Response::get_response( $cart_contents, $this->namespace, $this->rest_base );
+			return $this->get_response( $request, $cart_contents );
 		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
+			return $this->get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
 		}
 	} // END get_cart()
 
@@ -407,487 +346,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	} // END return_cart_contents()
 
 	/**
-	 * Validate the product ID or SKU ID.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access protected
-	 *
-	 * @since   1.0.0 Introduced.
-	 * @version 3.0.11
-	 *
-	 * @param int|string $product_id The product ID to validate.
-	 *
-	 * @return int $product_id The validated product ID.
-	 */
-	protected function validate_product_id( $product_id ) {
-		try {
-			// If the product ID was used by a SKU ID, then look up the product ID and return it.
-			if ( ! is_numeric( $product_id ) ) {
-				$product_id_by_sku = wc_get_product_id_by_sku( $product_id );
-
-				if ( ! empty( $product_id_by_sku ) && $product_id_by_sku > 0 ) {
-					$product_id = $product_id_by_sku;
-				} else {
-					$message = __( 'Product does not exist! Check that you have submitted a product ID or SKU ID correctly for a product that exists.', 'cart-rest-api-for-woocommerce' );
-
-					throw new CoCart_Data_Exception( 'cocart_unknown_product_id', $message, 404 );
-				}
-			}
-
-			if ( empty( $product_id ) ) {
-				$message = __( 'Product ID number is required!', 'cart-rest-api-for-woocommerce' );
-
-				throw new CoCart_Data_Exception( 'cocart_product_id_required', $message, 404 );
-			}
-
-			if ( ! is_numeric( $product_id ) ) {
-				$message = __( 'Product ID must be numeric!', 'cart-rest-api-for-woocommerce' );
-
-				throw new CoCart_Data_Exception( 'cocart_product_id_not_numeric', $message, 405 );
-			}
-
-			// Force product ID to be integer.
-			$product_id = (int) $product_id;
-
-			return $product_id;
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END validate_product_id()
-
-	/**
-	 * Validate the product quantity.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access protected
-	 *
-	 * @since 1.0.0 Introduced.
-	 * @since 3.1.0 Added product object as parameter and validation for maximum quantity allowed to add to cart.
-	 *
-	 * @param int|float  $quantity The quantity to validate.
-	 * @param WC_Product $product  Product object.
-	 */
-	protected function validate_quantity( $quantity, WC_Product $product = null ) {
-		try {
-			if ( ! is_numeric( $quantity ) ) {
-				throw new CoCart_Data_Exception( 'cocart_quantity_not_numeric', __( 'Quantity must be integer or a float value!', 'cart-rest-api-for-woocommerce' ), 405 );
-			}
-
-			/**
-			 * Filter allows control over the minimum quantity a customer must add to purchase said item.
-			 *
-			 * @since 3.0.17 Introduced.
-			 * @since 3.1.0 Added product object as parameter.
-			 *
-			 * @param int|float  Minimum quantity to validate with.
-			 * @param WC_Product Product object.
-			 */
-			$minimum_quantity = apply_filters( 'cocart_quantity_minimum_requirement', $product->get_min_purchase_quantity(), $product );
-
-			if ( 0 === $quantity || $quantity < $minimum_quantity ) {
-				throw new CoCart_Data_Exception( 'cocart_quantity_invalid_amount', sprintf(
-					/* translators: %s: Minimum quantity. */
-					__( 'Quantity must be a minimum of %s.', 'cart-rest-api-for-woocommerce' ),
-					$minimum_quantity
-				), 405 );
-			}
-
-			$maximum_quantity = ( ( $product->get_max_purchase_quantity() < 0 ) ) ? '' : $product->get_max_purchase_quantity(); // We replace -1 with a blank if stock management is not used.
-
-			/**
-			 * Filter allows control over the maximum quantity a customer
-			 * is able to add said item to the cart.
-			 *
-			 * @since 3.1.0 Introduced.
-			 *
-			 * @param int|float  Maximum quantity to validate with.
-			 * @param WC_Product Product object.
-			 */
-			$maximum_quantity = apply_filters( 'cocart_quantity_maximum_allowed', $maximum_quantity, $product );
-
-			if ( ! empty( $maximum_quantity ) && $quantity > $maximum_quantity ) {
-				throw new CoCart_Data_Exception( 'cocart_quantity_invalid_amount', sprintf(
-					/* translators: %s: Maximum quantity. */
-					__( 'Quantity must be %s or lower.', 'cart-rest-api-for-woocommerce' ),
-					$maximum_quantity
-				), 405 );
-			}
-
-			return wc_stock_amount( $quantity );
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END validate_quantity()
-
-	/**
-	 * Validate variable product.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access protected
-	 *
-	 * @since   2.1.0 Introduced.
-	 * @version 3.0.6
-	 *
-	 * @param int        $variation_id ID of the variation.
-	 * @param array      $variation    Attribute values.
-	 * @param WC_Product $product      Product object.
-	 *
-	 * @return array $variation_id ID of the variation and attribute values.
-	 */
-	protected function validate_variable_product( int $variation_id, array $variation, WC_Product $product ) {
-		try {
-			// Flatten data and format posted values.
-			$variable_product_attributes = $this->get_variable_product_attributes( $product );
-
-			// Now we have a variation ID, get the valid set of attributes for this variation. They will have an attribute_ prefix since they are from meta.
-			$expected_attributes = wc_get_product_variation_attributes( $variation_id );
-			$missing_attributes  = array();
-
-			foreach ( $variable_product_attributes as $attribute ) {
-				if ( ! $attribute['is_variation'] ) {
-					continue;
-				}
-
-				$prefixed_attribute_name = 'attribute_' . sanitize_title( $attribute['name'] );
-				$expected_value          = isset( $expected_attributes[ $prefixed_attribute_name ] ) ? $expected_attributes[ $prefixed_attribute_name ] : '';
-				$attribute_label         = wc_attribute_label( $attribute['name'] );
-
-				if ( isset( $variation[ wc_variation_attribute_name( $attribute['name'] ) ] ) ) {
-					$given_value = $variation[ wc_variation_attribute_name( $attribute['name'] ) ];
-
-					if ( $expected_value === $given_value ) {
-						continue;
-					}
-
-					// If valid values are empty, this is an 'any' variation so get all possible values.
-					if ( '' === $expected_value && in_array( $given_value, $attribute->get_slugs(), true ) ) {
-						continue;
-					}
-
-					/* translators: %1$s: Attribute name, %2$s: Allowed values. */
-					$message = sprintf( __( 'Invalid value posted for %1$s. Allowed values: %2$s', 'cart-rest-api-for-woocommerce' ), $attribute_label, implode( ', ', $attribute->get_slugs() ) );
-
-					/**
-					 * Filters message about invalid variation data.
-					 *
-					 * @param string $message         Message.
-					 * @param string $attribute_label Attribute Label.
-					 * @param array  $attribute       Allowed values.
-					 */
-					$message = apply_filters( 'cocart_invalid_variation_data_message', $message, $attribute_label, $attribute->get_slugs() );
-
-					throw new CoCart_Data_Exception( 'cocart_invalid_variation_data', $message, 400 );
-				}
-
-				// If no attribute was posted, only error if the variation has an 'any' attribute which requires a value.
-				if ( '' === $expected_value ) {
-					$missing_attributes[] = $attribute_label;
-				}
-			}
-
-			if ( ! empty( $missing_attributes ) ) {
-				/* translators: %s: Attribute name. */
-				$message = __( 'Missing variation data for variable product.', 'cart-rest-api-for-woocommerce' ) . ' ' . sprintf( _n( '%s is a required field.', '%s are required fields.', count( $missing_attributes ), 'cart-rest-api-for-woocommerce' ), wc_format_list_of_items( $missing_attributes ) );
-
-				/**
-				 * Filters message about missing variation data.
-				 *
-				 * @param string $message            Message.
-				 * @param string $missing_attributes Number of missing attributes.
-				 * @param array  $missing_attributes List of missing attributes.
-				 */
-				$message = apply_filters( 'cocart_missing_variation_data_message', $message, count( $missing_attributes ), wc_format_list_of_items( $missing_attributes ) );
-
-				throw new CoCart_Data_Exception( 'cocart_missing_variation_data', $message, 400 );
-			}
-
-			return $variation;
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END validate_variable_product()
-
-	/**
-	 * Tries to match variation attributes passed to a variation ID and return the ID.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access protected
-	 *
-	 * @since   2.1.2 Introduced.
-	 * @version 3.0.0
-	 *
-	 * @param array      $variation Submitted attributes.
-	 * @param WC_Product $product   Product object.
-	 *
-	 * @return int $variation_id Matching variation ID.
-	 */
-	protected function get_variation_id_from_variation_data( $variation, $product ) {
-		try {
-			$data_store   = \WC_Data_Store::load( 'product' );
-			$variation_id = $data_store->find_matching_product_variation( $product, $variation );
-
-			if ( empty( $variation_id ) ) {
-				$message = __( 'No matching variation found.', 'cart-rest-api-for-woocommerce' );
-
-				throw new CoCart_Data_Exception( 'cocart_no_variation_found', $message, 404 );
-			}
-
-			return $variation_id;
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END get_variation_id_from_variation_data()
-
-	/**
-	 * Get product attributes from the variable product (which may be the parent if the product object is a variation).
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access protected
-	 *
-	 * @since   2.1.2 Introduced.
-	 * @version 3.0.0
-	 *
-	 * @param WC_Product $product Product object.
-	 *
-	 * @return array $attributes Product attributes.
-	 */
-	protected function get_variable_product_attributes( $product ) {
-		try {
-			if ( $product->is_type( 'variation' ) ) {
-				$product = wc_get_product( $product->get_parent_id() );
-			}
-
-			if ( ! $product || 'trash' === $product->get_status() ) {
-				$message = __( 'This product cannot be added to the cart.', 'cart-rest-api-for-woocommerce' );
-
-				throw new CoCart_Data_Exception( 'cocart_cart_invalid_parent_product', $message, 404 );
-			}
-
-			return $product->get_attributes();
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END get_variable_product_attributes()
-
-	/**
-	 * Validate product before it is added to the cart, updated or removed.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access protected
-	 *
-	 * @since   1.0.0 Introduced.
-	 * @since   3.0.0 Deprecated $variation_id parameter is no longer used.
-	 * @version 3.1.0
-	 *
-	 * @param int             $product_id   Contains the ID of the product.
-	 * @param int|float       $quantity     Contains the quantity of the item.
-	 * @param null            $variation_id Used to pass the variation id of the product to add to the cart.
-	 * @param array           $variation    Contains the selected attributes.
-	 * @param array           $item_data    Extra cart item data we want to pass into the item.
-	 * @param string          $product_type The product type.
-	 * @param WP_REST_Request $request      Full details about the request.
-	 *
-	 * @return array Item data.
-	 */
-	protected function validate_product( $product_id = null, $quantity = 1, $variation_id = null, $variation = array(), $item_data = array(), $product_type = '', $request = array() ) {
-		try {
-			// Get product and validate product for the cart.
-			$product = wc_get_product( $product_id );
-			$product = $this->validate_product_for_cart( $product );
-
-			// Look up the product type if not passed.
-			if ( empty( $product_type ) ) {
-				$product_type = $product->get_type();
-			}
-
-			$variation_id = 0;
-
-			// Set correct product ID's if product type is a variation.
-			if ( $product->is_type( 'variation' ) ) {
-				$product_id   = $product->get_parent_id();
-				$variation_id = $product->get_id();
-			}
-
-			// If we have a parent product and no variation ID, find the variation ID.
-			if ( $product->is_type( 'variable' ) && 0 === $variation_id ) {
-				$variation_id = $this->get_variation_id_from_variation_data( $variation, $product );
-			}
-
-			// Throw exception if no variation is found.
-			if ( is_wp_error( $variation_id ) ) {
-				return $variation_id;
-			}
-
-			// Validate variable/variation product.
-			if ( 'variable' === $product_type || 'variation' === $product_type ) {
-				$variation = $this->validate_variable_product( $variation_id, $variation, $product );
-			}
-
-			/**
-			 * If variables are not valid then return error response.
-			 *
-			 * @param $variation
-			 */
-			if ( is_wp_error( $variation ) ) {
-				return $variation;
-			}
-
-			/**
-			 * Filters add to cart validation.
-			 *
-			 * @since 1.0.0 Introduced.
-			 *
-			 * @param bool   true          Default is true to allow the product to pass validation.
-			 * @param int    $product_id   Contains the ID of the product.
-			 * @param int    $quantity     Contains the quantity of the item.
-			 * @param int    $variation_id Used to pass the variation id of the product to add to the cart.
-			 * @param array  $variation    Contains the selected attributes.
-			 * @param object $item_data    Extra cart item data we want to pass into the item.
-			 * @param string $product_type The product type.
-			 */
-			$passed_validation = apply_filters( 'cocart_add_to_cart_validation', true, $product_id, $quantity, $variation_id, $variation, $item_data, $product_type, $request );
-
-			/**
-			 * If validation returned an error return error response.
-			 *
-			 * @param $passed_validation
-			 */
-			if ( is_wp_error( $passed_validation ) ) {
-				return $passed_validation;
-			}
-
-			// If validation returned false.
-			if ( ! $passed_validation ) {
-				$message = __( 'Product did not pass validation!', 'cart-rest-api-for-woocommerce' );
-
-				/**
-				 * Filters message about product failing validation.
-				 *
-				 * @since 1.0.0 Introduced.
-				 *
-				 * @param string     $message Message.
-				 * @param WC_Product $product Product object.
-				 */
-				$message = apply_filters( 'cocart_product_failed_validation_message', $message, $product );
-
-				throw new CoCart_Data_Exception( 'cocart_product_failed_validation', $message, 400 );
-			}
-
-			/**
-			 * Filter allows other plugins to add their own cart item data.
-			 *
-			 * @since 1.0.0 Introduced.
-			 *
-			 * @param array           $item_data    Extra cart item data we want to pass into the item.
-			 * @param int             $product_id   Contains the ID of the product.
-			 * @param null            $variation_id Used to pass the variation id of the product to add to the cart.
-			 * @param int|float       $quantity     Contains the quantity of the item.
-			 * @param string          $product_type The product type.
-			 * @param WP_REST_Request $request      Full details about the request.
-			 */
-			$item_data = (array) apply_filters( 'cocart_add_cart_item_data', $item_data, $product_id, $variation_id, $quantity, $product_type, $request );
-
-			// Generate an ID based on product ID, variation ID, variation data, and other cart item data.
-			$cart_id = $this->get_cart_instance()->generate_cart_id( $product_id, $variation_id, $variation, $item_data );
-
-			// Find the cart item key in the existing cart.
-			$item_key = $this->find_product_in_cart( $cart_id );
-
-			/**
-			 * Filters the quantity for specified products.
-			 *
-			 * @since 1.0.0 Introduced.
-			 *
-			 * @param int|float $quantity     The original quantity of the item.
-			 * @param int       $product_id   The product ID.
-			 * @param int       $variation_id The variation ID.
-			 * @param array     $variation    The variation data.
-			 * @param array     $item_data    The cart item data.
-			 */
-			$quantity = apply_filters( 'cocart_add_to_cart_quantity', $quantity, $product_id, $variation_id, $variation, $item_data );
-
-			// Validates the item quantity.
-			$quantity = $this->validate_item_quantity( $product, $quantity, $product_id, $variation_id, $item_data, $cart_id, $item_key );
-
-			/**
-			 * If product validation returned an error return error response.
-			 *
-			 * @param $quantity
-			 */
-			if ( is_wp_error( $quantity ) ) {
-				return $quantity;
-			}
-
-			// Validates the item before adding to cart.
-			$is_valid = $this->validate_add_to_cart( $product, $quantity );
-
-			/**
-			 * If product validation returned an error return error response.
-			 *
-			 * @param $is_valid
-			 */
-			if ( is_wp_error( $is_valid ) ) {
-				return $is_valid;
-			}
-
-			// Returns all valid data.
-			return array(
-				'product_id'   => $product_id,
-				'quantity'     => $quantity,
-				'variation_id' => $variation_id,
-				'variation'    => $variation,
-				'item_data'    => $item_data,
-				'item_key'     => $item_key,
-				'product_data' => $product,
-				'request'      => $request,
-			);
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END validate_product()
-
-	/**
-	 * Checks if the product in the cart has enough stock
-	 * before updating the quantity.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access protected
-	 *
-	 * @since   1.0.6 Introduced.
-	 * @version 3.0.0
-	 *
-	 * @param array     $current_data Cart item details.
-	 * @param int|float $quantity     The quantity to check stock.
-	 *
-	 * @return bool
-	 */
-	protected function has_enough_stock( $current_data = array(), $quantity = 1 ) {
-		try {
-			$product_id      = ! isset( $current_data['product_id'] ) ? 0 : absint( $current_data['product_id'] );
-			$variation_id    = ! isset( $current_data['variation_id'] ) ? 0 : absint( $current_data['variation_id'] );
-			$current_product = wc_get_product( $variation_id ? $variation_id : $product_id );
-
-			if ( ! $current_product->has_enough_stock( $quantity ) ) {
-				/* translators: 1: Quantity Requested, 2: Product Name 3: Quantity in Stock */
-				$message = sprintf( __( 'You cannot add a quantity of (%1$s) for "%2$s" to the cart because there is not enough stock. - only (%3$s remaining)!', 'cart-rest-api-for-woocommerce' ), $quantity, $current_product->get_name(), wc_format_stock_quantity_for_display( $current_product->get_stock_quantity(), $current_product ) );
-
-				throw new CoCart_Data_Exception( 'cocart_not_enough_in_stock', $message, 404 );
-			}
-
-			return true;
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END has_enough_stock()
-
-	/**
 	 * Prepares a list of store currency data to return in responses.
 	 *
 	 * @access public
@@ -904,29 +362,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 
 		return cocart_get_store_currency();
 	} // END get_store_currency()
-
-	/**
-	 * Returns the cart key.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access public
-	 *
-	 * @param WP_REST_Request $request Full details about the request.
-	 *
-	 * @return string Cart key.
-	 */
-	public function get_cart_key( $request ) {
-		try {
-			if ( ! method_exists( WC()->session, 'get_cart_key' ) ) {
-				return '';
-			}
-
-			return WC()->session->get_cart_key();
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END get_cart_key()
 
 	/**
 	 * Get tax lines from the cart and format to match schema.
@@ -1019,85 +454,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	} // END format_variation_data()
 
 	/**
-	 * Get cart fees.
-	 *
-	 * @access public
-	 *
-	 * @param WC_Cart $cart Cart class instance.
-	 *
-	 * @return array Cart fees.
-	 */
-	public function get_fees( $cart ) {
-		$cart_fees = $cart->get_fees();
-
-		$fees = array();
-
-		if ( ! empty( $cart_fees ) ) {
-			foreach ( $cart_fees as $key => $fee ) {
-				$fees[ $key ] = array(
-					'name' => esc_html( $fee->name ),
-					'fee'  => (float) cocart_prepare_money_response( $this->fee_html( $cart, $fee ) ),
-				);
-			}
-		}
-
-		return $fees;
-	} // END get_fees()
-
-	/**
-	 * Get cart totals.
-	 *
-	 * Returns the cart subtotal, fees, discounted total, shipping total
-	 * and total of the cart.
-	 *
-	 * @access public
-	 *
-	 * @since 4.0.0 Introduced.
-	 *
-	 * @param WP_REST_Request $request Full details about the request.
-	 * @param WC_Cart         $cart    Cart class instance.
-	 * @param array           $fields  An array of requested fields for the cart response to return.
-	 *
-	 * @return array Cart totals.
-	 */
-	public function get_cart_totals( $request = array(), $cart, $fields ) {
-		$totals = array(
-			'subtotal'       => $cart->get_subtotal(),
-			'subtotal_tax'   => $cart->get_subtotal_tax(),
-			'fee_total'      => $cart->get_fee_total(),
-			'fee_tax'        => $cart->get_fee_tax(),
-			'discount_total' => $cart->get_discount_total(),
-			'discount_tax'   => $cart->get_discount_tax(),
-			'shipping_total' => $cart->get_shipping_total(),
-			'shipping_tax'   => $cart->get_shipping_tax(),
-			'total'          => $cart->get_total( 'edit' ),
-			'total_tax'      => $cart->get_total_tax(),
-		);
-
-		if ( ! in_array( 'fees', $fields ) ) {
-			unset( $totals['fee_total'] );
-			unset( $totals['fee_tax'] );
-		}
-
-		if ( ! in_array( 'shipping', $fields ) ) {
-			unset( $totals['shipping_total'] );
-			unset( $totals['shipping_tax'] );
-		}
-
-		/**
-		 * Filters the cart totals.
-		 *
-		 * @since 4.0.0 Introduced.
-		 *
-		 * @param array           $totals  Cart totals.
-		 * @param WP_REST_Request $request Full details about the request.
-		 * @param WC_Cart         $cart    Cart class instance.
-		 * @param array           $fields  An array of requested fields for the cart response to return.
-		 */
-		return apply_filters( 'cocart_cart_totals', $totals, $request, $cart, $fields );
-	} // END get_cart_totals()
-
-	/**
 	 * Get coupon in HTML.
 	 *
 	 * @access public
@@ -1151,187 +507,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	} // END fee_html()
 
 	/**
-	 * Validates a product object for the cart.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access public
-	 *
-	 * @param WC_Product $product Product object.
-	 *
-	 * @return WC_Product $product Returns a product object if purchasable.
-	 */
-	public function validate_product_for_cart( $product ) {
-		try {
-			// Check if the product exists before continuing.
-			if ( ! $product || 'trash' === $product->get_status() ) {
-				$message = __( 'This product cannot be added to the cart.', 'cart-rest-api-for-woocommerce' );
-
-				/**
-				 * Filters message about product that cannot be added to cart.
-				 *
-				 * @param string     $message Message.
-				 * @param WC_Product $product Product object.
-				 */
-				$message = apply_filters( 'cocart_product_cannot_be_added_message', $message, $product );
-
-				throw new CoCart_Data_Exception( 'cocart_invalid_product', $message, 400 );
-			}
-
-			return $product;
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END get_product_for_cart()
-
-	/**
-	 * Validates item quantity and checks if sold individually.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access public
-	 *
-	 * @since   3.0.0 Introduced.
-	 * @version 3.1.0
-	 *
-	 * @param WC_Product $product      Product object.
-	 * @param int|float  $quantity     The quantity to validate.
-	 * @param int        $product_id   The product ID.
-	 * @param int        $variation_id The variation ID.
-	 * @param array      $item_data    The cart item data.
-	 * @param string     $cart_id      Generated ID based on item in cart.
-	 * @param string     $item_key     The item key of the cart item.
-	 *
-	 * @return float $quantity The quantity returned.
-	 */
-	public function validate_item_quantity( $product, $quantity, $product_id, $variation_id, $item_data, $cart_id, $item_key ) {
-		try {
-			// Force quantity to 1 if sold individually and check for existing item in cart.
-			if ( $product->is_sold_individually() ) {
-				/**
-				 * Quantity for sold individual products can be filtered.
-				 *
-				 * @since 2.0.13 Introduced.
-				 */
-				$quantity = apply_filters( 'cocart_add_to_cart_sold_individually_quantity', 1 );
-
-				$cart_contents = $this->get_cart();
-
-				$found_in_cart = apply_filters( 'cocart_add_to_cart_sold_individually_found_in_cart', $item_key && $cart_contents[ $item_key ]['quantity'] > 0, $product_id, $variation_id, $item_data, $cart_id );
-
-				if ( $found_in_cart ) {
-					/* translators: %s: Product Name */
-					$message = sprintf( __( "You cannot add another '%s' to your cart.", 'cart-rest-api-for-woocommerce' ), $product->get_name() );
-
-					/**
-					 * Filters message about product not being allowed to add another.
-					 *
-					 * @since 3.0.0 Introduced.
-					 *
-					 * @param string     $message Message.
-					 * @param WC_Product $product Product object.
-					 */
-					$message = apply_filters( 'cocart_product_can_not_add_another_message', $message, $product );
-
-					throw new CoCart_Data_Exception( 'cocart_product_sold_individually', $message, 403 );
-				}
-			}
-
-			return $quantity;
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END validate_item_quantity()
-
-	/**
-	 * Validates item and check for errors before added to cart.
-	 *
-	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
-	 *
-	 * @access public
-	 *
-	 * @since   2.1.0 Introduced.
-	 * @version 3.1.0
-	 *
-	 * @param WC_Product $product  Product object.
-	 * @param int|float  $quantity Quantity of product to validate availability.
-	 */
-	public function validate_add_to_cart( $product, $quantity ) {
-		try {
-			// Product is purchasable check.
-			if ( ! $product->is_purchasable() ) {
-				$this->throw_product_not_purchasable( $product );
-			}
-
-			// Stock check - only check if we're managing stock and backorders are not allowed.
-			if ( ! $product->is_in_stock() ) {
-				/* translators: %s: Product name */
-				$message = sprintf( __( 'You cannot add "%s" to the cart because the product is out of stock.', 'cart-rest-api-for-woocommerce' ), $product->get_name() );
-
-				/**
-				 * Filters message about product is out of stock.
-				 *
-				 * @since 2.1.0 Introduced.
-				 *
-				 * @param string     $message Message.
-				 * @param WC_Product $product Product object.
-				 */
-				$message = apply_filters( 'cocart_product_is_out_of_stock_message', $message, $product );
-
-				throw new CoCart_Data_Exception( 'cocart_product_out_of_stock', $message, 404 );
-			}
-
-			if ( ! $product->has_enough_stock( $quantity ) ) {
-				$stock_quantity = wc_format_stock_quantity_for_display( $product->get_stock_quantity(), $product );
-
-				if ( $stock_quantity > 0 ) {
-					/* translators: 1: Quantity Requested, 2: Product Name, 3: Quantity in Stock */
-					$message = sprintf( __( 'You cannot add the amount of %1$s for "%2$s" to the cart because there is not enough stock, only (%3$s) remaining.', 'cart-rest-api-for-woocommerce' ), $quantity, $product->get_name(), $stock_quantity );
-				} else {
-					/* translators: 1: Product Name */
-					$message = sprintf( __( 'You cannot add %1$s to the cart as it is no longer in stock.', 'cart-rest-api-for-woocommerce' ), $product->get_name() );
-				}
-
-				/**
-				 * Filters message about product not having enough stock.
-				 *
-				 * @since 3.1.0 Introduced.
-				 *
-				 * @param string     $message        Message.
-				 * @param WC_Product $product        Product object.
-				 * @param int        $stock_quantity Quantity remaining.
-				 */
-				$message = apply_filters( 'cocart_product_not_enough_stock_message', $message, $product, $stock_quantity );
-
-				throw new CoCart_Data_Exception( 'cocart_not_enough_in_stock', $message, 404 );
-			}
-
-			// Stock check - this time accounting for whats already in-cart and look up what's reserved.
-			if ( $product->managing_stock() && ! $product->backorders_allowed() ) {
-				$qty_remaining = $this->get_remaining_stock_for_product( $product );
-				$qty_in_cart   = $this->get_product_quantity_in_cart( $product );
-
-				if ( $qty_remaining < $qty_in_cart + $quantity ) {
-					$message = sprintf(
-						/* translators: 1: product name, 2: Quantity in Stock, 3: Quantity in Cart */
-						__( 'You cannot add that amount of "%1$s" to the cart &mdash; we have (%2$s) in stock remaining. You already have (%3$s) in your cart.', 'cart-rest-api-for-woocommerce' ),
-						$product->get_name(),
-						wc_format_stock_quantity_for_display( $product->get_stock_quantity(), $product ),
-						wc_format_stock_quantity_for_display( $qty_in_cart, $product )
-					);
-
-					throw new CoCart_Data_Exception( 'cocart_not_enough_stock_remaining', $message, 404 );
-				}
-			}
-
-			cocart_deprecated_hook( 'cocart_ok_to_add_response', '3.0.0', null, 'This filter is no longer used in the API.' );
-			cocart_deprecated_hook( 'cocart_ok_to_add', '3.0.0', null, 'This filter is no longer used in the API.' );
-		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
-		}
-	} // END validate_add_to_cart()
-
-	/**
 	 * Filters additional requested data.
 	 *
 	 * @access public
@@ -1371,53 +546,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 
 		return $product_slug;
 	} // END get_product_slug()
-
-	/**
-	 * Returns the item thumbnail from the requested product in cart.
-	 *
-	 * @access public
-	 *
-	 * @since 4.0.0 Introduced.
-	 *
-	 * @param WC_Product $product      Product object.
-	 * @param array      $cart_item    The item in the cart containing the default cart item data.
-	 * @param string     $item_key     The item key generated based on the details of the item.
-	 * @param boolean    $removed_item Determines if the item in the cart is removed.
-	 *
-	 * @return string
-	 */
-	public function get_item_thumbnail( $product, $cart_item, $item_key, $removed_item ) {
-		$thumbnail_id = ! empty( $product->get_image_id() ) ? $product->get_image_id() : get_option( 'woocommerce_placeholder_image', 0 );
-
-		/**
-		 * Filter allows you to change the thumbnail ID of the item in cart.
-		 *
-		 * @since 3.0.0 Introduced.
-		 *
-		 * @param int     $thumbnail_id Thumbnail ID of the image.
-		 * @param array   $cart_item    The item in the cart containing the default cart item data.
-		 * @param string  $item_key     The item key generated based on the details of the item.
-		 * @param boolean $removed_item Determines if the item in the cart is removed.
-		 */
-		$thumbnail_id = apply_filters( 'cocart_item_thumbnail', $thumbnail_id, $cart_item, $item_key, $removed_item );
-
-		$thumbnail_src = wp_get_attachment_image_src( $thumbnail_id, apply_filters( 'cocart_item_thumbnail_size', 'woocommerce_thumbnail', $removed_item ) );
-
-		$thumbnail_src = ! empty( $thumbnail_src[0] ) ? $thumbnail_src[0] : '';
-
-		/**
-		 * Filters the source of the product thumbnail of the item in cart.
-		 *
-		 * @since   2.1.0 Introduced.
-		 * @version 3.0.0
-		 *
-		 * @param string $thumbnail_src URL of the product thumbnail.
-		 */
-		$thumbnail_src = apply_filters( 'cocart_item_thumbnail_src', $thumbnail_src, $cart_item, $item_key, $removed_item );
-
-		// Return main featured image.
-		return esc_url( $thumbnail_src );
-	} // END get_item_thumbnail()
 
 	/**
 	 * Get a single item from the cart and present the data required.
@@ -1628,32 +756,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	} // END get_removed_items()
 
 	/**
-	 * Removes all internal elements of an item that is not needed.
-	 *
-	 * @access protected
-	 *
-	 * @param array $cart_item Before cart item data is modified.
-	 *
-	 * @return array $cart_item Modified cart item data returned.
-	 */
-	protected function prepare_item( $cart_item ) {
-		unset( $cart_item['key'] );
-		unset( $cart_item['product_id'] );
-		unset( $cart_item['variation_id'] );
-		unset( $cart_item['variation'] );
-		unset( $cart_item['quantity'] );
-		unset( $cart_item['data'] );
-		unset( $cart_item['data_hash'] );
-		unset( $cart_item['line_tax_data'] );
-		unset( $cart_item['line_subtotal'] );
-		unset( $cart_item['line_subtotal_tax'] );
-		unset( $cart_item['line_total'] );
-		unset( $cart_item['line_tax'] );
-
-		return $cart_item;
-	} // END prepare_item()
-
-	/**
 	 * Returns cross sells based on the items in the cart.
 	 *
 	 * @access public
@@ -1805,78 +907,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	} // END get_shipping_details()
 
 	/**
-	 * Cleans up the meta data for API.
-	 *
-	 * @access protected
-	 *
-	 * @since   3.1.0 Introduced
-	 * @version 3.1.2
-	 *
-	 * @param object $method Method data.
-	 * @param string $type   Meta data we are cleaning for.
-	 *
-	 * @return array Meta data.
-	 */
-	protected function clean_meta_data( $method, $type = 'shipping' ) {
-		$meta_data = $method->get_meta_data();
-
-		switch ( $type ) {
-			case 'shipping':
-				$meta_data['items'] = isset( $meta_data['Items'] ) ? html_entity_decode( wp_strip_all_tags( $meta_data['Items'] ) ) : '';
-				unset( $meta_data['Items'] );
-
-				break;
-		}
-
-		return $meta_data;
-	} // END clean_meta_data()
-
-	/**
-	 * Return notices in cart if any.
-	 *
-	 * @access protected
-	 *
-	 * @return array $notices.
-	 */
-	protected function maybe_return_notices() {
-		$notice_count = 0;
-		$all_notices  = WC()->session->get( 'wc_notices', array() );
-
-		foreach ( $all_notices as $notices ) {
-			$notice_count += count( $notices );
-		}
-
-		$notices = $notice_count > 0 ? $this->print_notices() : array();
-
-		return $notices;
-	} // END maybe_return_notices()
-
-	/**
-	 * Returns messages and errors which are stored in the session, then clears them.
-	 *
-	 * @access protected
-	 *
-	 * @return array
-	 */
-	protected function print_notices() {
-		$all_notices  = WC()->session->get( 'wc_notices', array() );
-		$notice_types = apply_filters( 'cocart_notice_types', array( 'error', 'success', 'notice', 'info' ) );
-		$notices      = array();
-
-		foreach ( $notice_types as $notice_type ) {
-			if ( wc_notice_count( $notice_type ) > 0 ) {
-				foreach ( $all_notices[ $notice_type ] as $key => $notice ) {
-					$notices[ $notice_type ][ $key ] = wc_kses_notice( $notice['notice'] );
-				}
-			}
-		}
-
-		wc_clear_notices();
-
-		return $notices;
-	} // END print_notices()
-
-	/**
 	 * Adds item to cart internally rather than WC.
 	 *
 	 * @throws CoCart_Data_Exception Exception if invalid data is detected.
@@ -1924,26 +954,9 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 
 			return $item_key;
 		} catch ( CoCart_Data_Exception $e ) {
-			return CoCart_Response::get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
+			return $this->get_error_response( $e->getErrorCode(), $e->getMessage(), $e->getCode(), $e->getAdditionalData() );
 		}
 	} // END add_cart_item()
-
-	/**
-	 * Cache cart item.
-	 *
-	 * @see CartCache::set_cached_item()
-	 *
-	 * @access public
-	 *
-	 * @since 3.1.0 Introduced.
-	 *
-	 * @param array $was_added_to_cart Cart item to cache.
-	 */
-	public function cache_cart_item( $was_added_to_cart ) {
-		$item_key = $was_added_to_cart['key'];
-
-		CartCache::set_cached_item( $item_key, $was_added_to_cart );
-	} // END cache_cart_item()
 
 	/**
 	 * Returns the customers details from fields.
@@ -1987,37 +1000,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 
 		return $results;
 	} // END get_customer_fields()
-
-	/**
-	 * Convert queued error notices into an exception.
-	 *
-	 * For example, Payment methods may add error notices during validating fields to prevent checkout.
-	 * This method will find the first error message and thrown an exception instead. Discards notices once complete.
-	 *
-	 * @throws CoCart_Data_Exception If an error notice is detected, Exception is thrown.
-	 *
-	 * @access public
-	 *
-	 * @static
-	 *
-	 * @since 3.0.1 Introduced.
-	 *
-	 * @param string $error_code Error code for the thrown exceptions.
-	 */
-	public static function convert_notices_to_exceptions( $error_code = 'unknown_server_error' ) {
-		if ( 0 === wc_notice_count( 'error' ) ) {
-			return;
-		}
-
-		$error_notices = wc_get_notices( 'error' );
-
-		// Prevent notices from being output later on.
-		wc_clear_notices();
-
-		foreach ( $error_notices as $error_notice ) {
-			throw new CoCart_Data_Exception( $error_code, wp_strip_all_tags( $error_notice['notice'] ), 400 );
-		}
-	} // END convert_notices_to_exceptions()
 
 	/**
 	 * Get cart template.
@@ -2231,411 +1213,6 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	} // END get_cart_template_limited()
 
 	/**
-	 * Throws exception when an item cannot be added to the cart.
-	 *
-	 * @throws CoCart_Data_Exception If an error notice is detected, Exception is thrown.
-	 *
-	 * @access protected
-	 *
-	 * @since 3.0.4 Introduced.
-	 *
-	 * @param WC_Product $product Product object.
-	 */
-	protected function throw_product_not_purchasable( $product ) {
-		$message = sprintf(
-			/* translators: %s: product name */
-			__( "'%s' is not available for purchase.", 'cart-rest-api-for-woocommerce' ),
-			$product->get_name()
-		);
-
-		/**
-		 * Filters message about product unable to be purchased.
-		 *
-		 * @param string     $message Message.
-		 * @param WC_Product $product Product object.
-		 */
-		$message = apply_filters( 'cocart_product_cannot_be_purchased_message', $message, $product );
-
-		throw new CoCart_Data_Exception( 'cocart_cannot_be_purchased', $message, 400 );
-	} // END throw_product_not_purchasable()
-
-	/**
-	 * Gets the quantity of a product across line items.
-	 *
-	 * @access protected
-	 *
-	 * @since 3.1.0 Introduced.
-	 *
-	 * @param WC_Product $product Product object.
-	 *
-	 * @return int Quantity of the product.
-	 */
-	protected function get_product_quantity_in_cart( $product ) {
-		$cart               = $this->get_cart_instance();
-		$product_quantities = $cart->get_cart_item_quantities();
-		$product_id         = $product->get_stock_managed_by_id();
-
-		return isset( $product_quantities[ $product_id ] ) ? $product_quantities[ $product_id ] : 0;
-	} // END get_product_quantity_in_cart()
-
-	/**
-	 * Gets remaining stock for a product.
-	 *
-	 * @access protected
-	 *
-	 * @since 3.1.0 Introduced.
-	 *
-	 * @param WC_Product $product Product object.
-	 *
-	 * @return int Remaining stock.
-	 */
-	protected function get_remaining_stock_for_product( $product ) {
-		$reserve_stock = new ReserveStock();
-		$draft_order   = WC()->session->get( 'cocart_draft_order', 0 );
-		$qty_reserved  = $reserve_stock->get_reserved_stock( $product, $draft_order );
-
-		return $product->get_stock_quantity() - $qty_reserved;
-	} // END get_remaining_stock_for_product()
-
-	/**
-	 * Throws exception if the item key is not provided when either removing, updating or restoring the item.
-	 *
-	 * @throws CoCart_Data_Exception If an error notice is detected, Exception is thrown.
-	 *
-	 * @access protected
-	 *
-	 * @since 3.0.17 Introduced.
-	 *
-	 * @param string $item_key Item key of the item in the cart.
-	 * @param string $status   Status of which we are checking the item key.
-	 *
-	 * @return string $item_key Item key of the item in the cart.
-	 */
-	protected function throw_missing_item_key( $item_key, $status ) {
-		$item_key = (string) $item_key; // Make sure the item key is a string value.
-
-		if ( '0' === $item_key ) {
-			$message = __( 'Missing cart item key is required!', 'cart-rest-api-for-woocommerce' );
-
-			/**
-			 * Filters message about cart item key required.
-			 *
-			 * @since 2.1.0 Introduced.
-			 *
-			 * @param string $message Message.
-			 */
-			$message = apply_filters( 'cocart_cart_item_key_required_message', $message, $status );
-
-			throw new CoCart_Data_Exception( 'cocart_cart_item_key_required', $message, 404 );
-		}
-
-		return $item_key;
-	} // END throw_missing_item_key()
-
-	/**
-	 * Ensures the cart totals are calculated before an API response is returned.
-	 *
-	 * @access public
-	 *
-	 * @since 3.1.0 Introduced.
-	 */
-	public function calculate_totals() {
-		$this->get_cart_instance()->calculate_fees();
-		$this->get_cart_instance()->calculate_shipping();
-		$this->get_cart_instance()->calculate_totals();
-	} // END calculate_totals()
-
-	/**
-	 * Returns an array of fields based on the configuration requested.
-	 *
-	 * @access public
-	 *
-	 * @since 4.0.0 Introduced.
-	 *
-	 * @param WP_REST_Request $request Full details about the request.
-	 *
-	 * @return array
-	 */
-	public function get_fields_configuration( $request ) {
-		$config = trim( $request['config']['fields'] );
-
-		switch ( $config ) {
-			case 'digital':
-				$fields = array( 'currency', 'customer', 'items', 'coupons', 'needs_payment', 'taxes', 'totals', 'notices' );
-				break;
-			case 'digital_fees':
-				$fields = array( 'currency', 'customer', 'items', 'coupons', 'needs_payment', 'fees', 'taxes', 'totals', 'notices' );
-				break;
-			case 'shipping':
-				$fields = array( 'currency', 'customer', 'items', 'items_weight', 'coupons', 'needs_payment', 'needs_shipping', 'shipping', 'taxes', 'totals', 'notices' );
-				break;
-			case 'shipping_fees':
-				$fields = array( 'currency', 'customer', 'items', 'items_weight', 'coupons', 'needs_payment', 'needs_shipping', 'shipping', 'fees', 'taxes', 'totals', 'notices' );
-				break;
-			case 'removed_items':
-				$fields = array( 'currency', 'removed_items', 'notices' );
-				break;
-			case 'cross_sells':
-				$fields = array( 'currency', 'cross_sells', 'notices' );
-				break;
-			default:
-				$fields = $this->get_fields_for_response( $request );
-				break;
-		}
-
-		return $fields;
-	} // END get_fields_configuration()
-
-	/**
-	 * Gets an array of fields to be included on the response.
-	 *
-	 * Included fields are based on item schema and `fields=` request argument.
-	 *
-	 * @access public
-	 *
-	 * @since 4.0.0 Introduced.
-	 *
-	 * @param WP_REST_Request $request Full details about the request.
-	 *
-	 * @return string Fields to be included in the response.
-	 */
-	public function get_fields_for_response( $request ) {
-		$schema     = $this->get_public_item_schema();
-		$properties = isset( $schema['properties'] ) ? $schema['properties'] : array();
-
-		$fields = array_unique( array_keys( $properties ) );
-
-		if ( ! isset( $request['fields'] ) ) {
-			return $fields;
-		}
-
-		$requested_fields = wp_parse_list( $request['fields'] );
-
-		// Return all fields if no fields specified.
-		if ( 0 === count( $requested_fields ) ) {
-			return $fields;
-		}
-
-		// Trim off outside whitespace from the comma delimited list.
-		$requested_fields = array_map( 'trim', $requested_fields );
-
-		// Return the list of all requested fields which appear in the schema.
-		return array_reduce(
-			$requested_fields,
-			static function( $response_fields, $field ) use ( $fields ) {
-				if ( in_array( $field, $fields, true ) ) {
-					$response_fields[] = $field;
-
-					return $response_fields;
-				}
-
-				// Check for nested fields if $field is not a direct match.
-				$nested_fields = explode( ':', $field );
-
-				// A nested field is included so long as its top-level property
-				// is present in the schema.
-				if ( in_array( $nested_fields[0], $fields, true ) ) {
-					$response_fields[] = $field;
-				}
-
-				return $response_fields;
-			},
-			array()
-		);
-	} // END get_fields_for_response()
-
-	/**
-	 * Retrieves the cart items schema properties.
-	 *
-	 * @access public
-	 *
-	 * @since 4.0.0 Introduced.
-	 *
-	 * @return array Public cart items schema properties.
-	 */
-	public function get_public_cart_items_schema_properties() {
-		return array(
-			'item_key'       => array(
-				'description' => __( 'Unique ID of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'string',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-			),
-			'id'             => array(
-				'description' => __( 'Product ID or Variation ID of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'integer',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-			),
-			'name'           => array(
-				'description' => __( 'The name of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'string',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-			),
-			'title'          => array(
-				'description' => __( 'The title of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'string',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-			),
-			'price'          => array(
-				'description' => __( 'The price of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'float',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-			),
-			'quantity'       => array(
-				'description' => __( 'The quantity of the item in the cart and minimum and maximum purchase capability.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'object',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-				'properties'  => array(
-					'value'        => array(
-						'description' => __( 'The quantity of the item.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'float',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-					'min_purchase' => array(
-						'description' => __( 'The minimum purchase amount required.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'float',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-					'max_purchase' => array(
-						'description' => __( 'The maximum purchase amount allowed. If -1 the item has an unlimited purchase amount.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'float',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-				),
-			),
-			'totals'         => array(
-				'description' => __( 'The totals of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'object',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-				'properties'  => array(
-					'subtotal'     => array(
-						'description' => __( 'The subtotal of the item.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'float',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-					'subtotal_tax' => array(
-						'description' => __( 'The subtotal tax of the item.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'float',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-					'total'        => array(
-						'description' => __( 'The total of the item.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'float',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-					'total_tax'    => array(
-						'description' => __( 'The total tax of the item.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'float',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-				),
-			),
-			'slug'           => array(
-				'description' => __( 'The product slug of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'string',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-			),
-			'meta'           => array(
-				'description' => __( 'The meta data of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'object',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-				'properties'  => array(
-					'product_type' => array(
-						'description' => __( 'The product type of the item.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'string',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-					'sku'          => array(
-						'description' => __( 'The SKU of the item.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'string',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-					'dimensions'   => array(
-						'description' => __( 'The dimensions of the item.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'object',
-						'context'     => array( 'view' ),
-						'properties'  => array(
-							'length' => array(
-								'description' => __( 'The length of the item.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-							'width'  => array(
-								'description' => __( 'The width of the item.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-							'height' => array(
-								'description' => __( 'The height of the item.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-							'unit'   => array(
-								'description' => __( 'The unit measurement of the item.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-						),
-					),
-					'weight'       => array(
-						'description' => __( 'The weight of the item.', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'float',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-					),
-					'variation'    => array(
-						'description' => __( 'The variation attributes of the item (if item is a variation of a variable product).', 'cart-rest-api-for-woocommerce' ),
-						'type'        => 'object',
-						'context'     => array( 'view' ),
-						'readonly'    => true,
-						'properties'  => array(),
-					),
-				),
-			),
-			'backorders'     => array(
-				'description' => __( 'The price of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'string',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-			),
-			'cart_item_data' => array(
-				'description' => __( 'Custom item data applied to the item (if any).', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'object',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-				'properties'  => array(),
-			),
-			'featured_image' => array(
-				'description' => __( 'The featured image of the item.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'string',
-				'context'     => array( 'view' ),
-				'readonly'    => true,
-			),
-		);
-	} // END get_public_cart_items_schema_properties()
-
-	/**
 	 * Retrieves the item schema for returning the cart.
 	 *
 	 * @access public
@@ -2645,669 +1222,7 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 	 * @return array Public item schema data.
 	 */
 	public function get_public_item_schema() {
-		$schema = array(
-			'$schema'    => 'http://json-schema.org/draft-04/schema#',
-			'title'      => 'cocart_cart',
-			'type'       => 'object',
-			'properties' => array(
-				'cart_hash'      => array(
-					'description' => __( 'A Unique hash key of the carts contents.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'string',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-				),
-				'cart_key'       => array(
-					'description' => __( 'The cart key identifying the cart in session.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'string',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-				),
-				'currency'       => array(
-					'description' => __( 'The store currency information.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'object',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'properties'  => array(
-						'currency_code'               => array(
-							'description' => __( 'The currency code (in ISO format) for returned prices.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'currency_symbol'             => array(
-							'description' => __( 'The currency symbol for the currency which can be used to format returned prices.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'currency_minor_unit'         => array(
-							'description' => __( 'The currency minor unit (number of digits after the decimal separator) for returned prices.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'integer',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'currency_decimal_separator'  => array(
-							'description' => __( 'The decimal separator for the currency which can be used to format returned prices.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'currency_thousand_separator' => array(
-							'description' => __( 'The thousand separator for the currency which can be used to format returned prices.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'currency_prefix'             => array(
-							'description' => __( 'The price prefix for the currency which can be used to format returned prices.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'currency_suffix'             => array(
-							'description' => __( 'The price prefix for the currency which can be used to format returned prices.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-					),
-				),
-				'customer'       => array(
-					'description' => __( 'The customer information.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'object',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'properties'  => array(
-						'billing_address'  => array(
-							'description' => __( 'Customers billing address.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'object',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-							'properties'  => array(
-								'billing_first_name' => array(
-									'description' => __( 'Customers billing first name.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_last_name'  => array(
-									'description' => __( 'Customers billing last name.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_company'    => array(
-									'description' => __( 'Customers billing company name.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_country'    => array(
-									'description' => __( 'Customers billing country.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_address_1'  => array(
-									'description' => __( 'Customers billing address line one.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_address_2'  => array(
-									'description' => __( 'Customers billing address line two.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_city'       => array(
-									'description' => __( 'Customers billing address city.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_state'      => array(
-									'description' => __( 'Customers billing state.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_postcode'   => array(
-									'description' => __( 'Customers billing postcode or zip code.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_phone'      => array(
-									'description' => __( 'Customers billing phone.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'billing_email'      => array(
-									'description' => __( 'Customers billing email address.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-							),
-						),
-						'shipping_address' => array(
-							'description' => __( 'Customers shipping address.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'object',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-							'properties'  => array(
-								'shipping_first_name' => array(
-									'description' => __( 'Customers shipping first name.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'shipping_last_name'  => array(
-									'description' => __( 'Customers shipping last name.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'shipping_company'    => array(
-									'description' => __( 'Customers shipping company name.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'shipping_country'    => array(
-									'description' => __( 'Customers shipping country.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'shipping_address_1'  => array(
-									'description' => __( 'Customers shipping address line one.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'shipping_address_2'  => array(
-									'description' => __( 'Customers shipping address line two.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'shipping_city'       => array(
-									'description' => __( 'Customers shipping address city.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'shipping_state'      => array(
-									'description' => __( 'Customers shipping state.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'shipping_postcode'   => array(
-									'description' => __( 'Customers shipping postcode or zip code.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-							),
-						),
-					),
-				),
-				'items'          => array(
-					'description' => __( 'The list of items in the cart.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'array',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'items'       => array(
-						'type'       => 'object',
-						'properties' => $this->get_public_cart_items_schema_properties(),
-					),
-				),
-				'item_count'     => array(
-					'description' => __( 'The number of items in the cart.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'integer',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-				),
-				'items_weight'   => array(
-					'description' => __( 'The total weight of all items in the cart.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'integer',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-				),
-				'coupons'        => array(
-					'description' => __( 'Coupons added to the cart.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'array',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'items'       => array(
-						'type'       => 'object',
-						'properties' => array(
-							'coupon'      => array(
-								'description' => __( 'The coupon code.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-							'label'       => array(
-								'description' => __( 'The coupon label.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-							'saving'      => array(
-								'description' => __( 'The amount discounted from the cart.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-							'saving_html' => array(
-								'description' => __( 'The amount discounted from the cart (HTML formatted).', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-						),
-					),
-				),
-				'needs_payment'  => array(
-					'description' => __( 'True if the cart needs payment. False for carts with only free products and no shipping costs.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'boolean',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-				),
-				'needs_shipping' => array(
-					'description' => __( 'True if the cart needs shipping and requires the showing of shipping costs.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'boolean',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-				),
-				'shipping'       => array(
-					'description' => __( 'The shipping packages available to the customer.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'object',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'properties'  => array(
-						'total_packages'          => array(
-							'description' => __( 'The number of shipping packages available calculated on the shipping address.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'integer',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'show_package_details'    => array(
-							'description' => __( 'True if the cart meets the criteria for showing items in the cart assigned to package.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'boolean',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'has_calculated_shipping' => array(
-							'description' => __( 'True if the cart meets the criteria for showing shipping costs, and rates have been calculated and included in the totals.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'boolean',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'packages'                => array(
-							'description' => __( 'The packages returned after calculating shipping.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'object',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-							'properties'  => array(
-								'default' => array(
-									'description' => __( 'The default package.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'object',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-									'properties'  => array(
-										'package_name'    => array(
-											'description' => __( 'The package name.', 'cart-rest-api-for-woocommerce' ),
-											'type'        => 'string',
-											'context'     => array( 'view' ),
-											'readonly'    => true,
-										),
-										'rates'           => array(
-											'description' => __( 'The packages returned after calculating shipping.', 'cart-rest-api-for-woocommerce' ),
-											'type'        => 'object',
-											'context'     => array( 'view' ),
-											'readonly'    => true,
-											'properties'  => array(
-												'[a-z0-9]' => array(
-													'key'  => array(
-														'description' => __( 'The rate key.', 'cart-rest-api-for-woocommerce' ),
-														'type'        => 'string',
-														'context'     => array( 'view' ),
-														'readonly'    => true,
-													),
-													'method_id' => array(
-														'description' => __( 'The method ID.', 'cart-rest-api-for-woocommerce' ),
-														'type'        => 'string',
-														'context'     => array( 'view' ),
-														'readonly'    => true,
-													),
-													'instance_id' => array(
-														'description' => __( 'The instance ID.', 'cart-rest-api-for-woocommerce' ),
-														'type'        => 'string',
-														'context'     => array( 'view' ),
-														'readonly'    => true,
-													),
-													'label' => array(
-														'description' => __( 'The rate label.', 'cart-rest-api-for-woocommerce' ),
-														'type'        => 'string',
-														'context'     => array( 'view' ),
-														'readonly'    => true,
-													),
-													'cost' => array(
-														'description' => __( 'The rate cost.', 'cart-rest-api-for-woocommerce' ),
-														'type'        => 'string',
-														'context'     => array( 'view' ),
-														'readonly'    => true,
-													),
-													'html' => array(
-														'description' => __( 'The rate label and cost formatted.', 'cart-rest-api-for-woocommerce' ),
-														'type'        => 'string',
-														'context'     => array( 'view' ),
-														'readonly'    => true,
-													),
-													'taxes' => array(
-														'description' => __( 'The rate tax cost.', 'cart-rest-api-for-woocommerce' ),
-														'type'        => 'string',
-														'context'     => array( 'view' ),
-														'readonly'    => true,
-													),
-													'chosen_method' => array(
-														'description' => __( 'The chosen method.', 'cart-rest-api-for-woocommerce' ),
-														'type'        => 'boolean',
-														'context'     => array( 'view' ),
-														'readonly'    => true,
-													),
-													'meta_data' => array(
-														'description' => __( 'The rate meta data.', 'cart-rest-api-for-woocommerce' ),
-														'type'        => 'object',
-														'context'     => array( 'view' ),
-														'readonly'    => true,
-														'properties' => array(
-															'items' => array(
-																'description' => __( 'The items the shipping rate has calculated based on.', 'cart-rest-api-for-woocommerce' ),
-																'type'        => 'string',
-																'context'     => array( 'view' ),
-																'readonly'    => true,
-															),
-														),
-													),
-												),
-											),
-										),
-										'package_details' => array(
-											'description' => __( 'The package details if any.', 'cart-rest-api-for-woocommerce' ),
-											'type'        => 'string',
-											'context'     => array( 'view' ),
-											'readonly'    => true,
-										),
-										'index'           => array(
-											'description' => __( 'The package index.', 'cart-rest-api-for-woocommerce' ),
-											'type'        => 'integer',
-											'context'     => array( 'view' ),
-											'readonly'    => true,
-										),
-										'chosen_method'   => array(
-											'description' => __( 'The chosen method set from the available rates.', 'cart-rest-api-for-woocommerce' ),
-											'type'        => 'string',
-											'context'     => array( 'view' ),
-											'readonly'    => true,
-										),
-										'formatted_destination' => array(
-											'description' => __( 'The destination the package.', 'cart-rest-api-for-woocommerce' ),
-											'type'        => 'string',
-											'context'     => array( 'view' ),
-											'readonly'    => true,
-										),
-									),
-								),
-							),
-						),
-					),
-				),
-				'fees'           => array(
-					'description' => __( 'The cart fees.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'object',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'properties'  => array(
-						'[a-zA-Z0-9]' => array(
-							'name' => array(
-								'description' => __( 'The fee name.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-							'fee'  => array(
-								'description' => __( 'The fee value.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'string',
-								'context'     => array( 'view' ),
-								'readonly'    => true,
-							),
-						),
-					),
-				),
-				'taxes'          => array(
-					'description' => __( 'The cart taxes.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'object',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'properties'  => array(
-						'[A-Z-TAX-0-9]' => array(
-							'description' => __( 'The store currency information.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'object',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-							'properties'  => array(
-								'name'  => array(
-									'description' => __( 'The tax name.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-								'price' => array(
-									'description' => __( 'The tax price.', 'cart-rest-api-for-woocommerce' ),
-									'type'        => 'string',
-									'context'     => array( 'view' ),
-									'readonly'    => true,
-								),
-							),
-						),
-					),
-				),
-				'totals'         => array(
-					'description' => __( 'The store currency information.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'object',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'properties'  => array(
-						'subtotal'       => array(
-							'description' => __( 'The subtotal of all items, shipping (if any) and fees (if any) before coupons applied (if any) to the cart.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'subtotal_tax'   => array(
-							'description' => __( 'The subtotal tax of all items, shipping (if any) and fees (if any) before coupons applied (if any) to the cart.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'fee_total'      => array(
-							'description' => __( 'The fee total.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'fee_tax'        => array(
-							'description' => __( 'The fee tax.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'discount_total' => array(
-							'description' => __( 'The discount total.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'discount_tax'   => array(
-							'description' => __( 'The discount tax.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'shipping_total' => array(
-							'description' => __( 'The shipping total of the selected packages.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'shipping_tax'   => array(
-							'description' => __( 'The shipping tax.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'total'          => array(
-							'description' => __( 'The total of everything in the cart.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'total_tax'      => array(
-							'description' => __( 'The total tax.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-					),
-				),
-				'removed_items'  => array(
-					'description' => __( 'Items that have been removed from the cart.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'array',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'items'       => array(
-						'type'       => 'object',
-						'properties' => $this->get_public_cart_items_schema_properties(),
-					),
-				),
-				'cross_sells'    => array(
-					'description' => __( 'Items you may be interested in adding to the cart.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'array',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'properties'  => array(
-						'id'             => array(
-							'description' => __( 'Product ID or Variation ID of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'integer',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'name'           => array(
-							'description' => __( 'The name of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'title'          => array(
-							'description' => __( 'The title of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'slug'           => array(
-							'description' => __( 'The slug of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'price'          => array(
-							'description' => __( 'The price of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'regular_price'  => array(
-							'description' => __( 'The regular price of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'sale_price'     => array(
-							'description' => __( 'The sale price of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'image'          => array(
-							'description' => __( 'The image of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'average_rating' => array(
-							'description' => __( 'The average rating of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'on_sale'        => array(
-							'description' => __( 'The sale status of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'boolean',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-						'type'           => array(
-							'description' => __( 'The product type of the cross-sell product.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'string',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-						),
-					),
-				),
-				'notices'        => array(
-					'description' => __( 'The cart notices.', 'cart-rest-api-for-woocommerce' ),
-					'type'        => 'object',
-					'context'     => array( 'view' ),
-					'readonly'    => true,
-					'properties'  => array(
-						'success' => array(
-							'description' => __( 'Notices for successful actions.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'array',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-							'items'       => array(
-								'type' => 'string',
-							),
-						),
-						'info'    => array(
-							'description' => __( 'Notices for informational actions.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'array',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-							'items'       => array(
-								'type' => 'string',
-							),
-						),
-						'error'   => array(
-							'description' => __( 'Notices for error actions.', 'cart-rest-api-for-woocommerce' ),
-							'type'        => 'array',
-							'context'     => array( 'view' ),
-							'readonly'    => true,
-							'items'       => array(
-								'type' => 'string',
-							),
-						),
-					),
-				),
-			),
-		);
+		$schema = parent::get_public_item_schema();
 
 		/**
 		 * This filter is now deprecated and is replaced with `cocart_cart_items_schema`.
@@ -3329,90 +1244,5 @@ class CoCart_REST_Cart_v2_Controller extends CoCart_API_Controller {
 
 		return $schema;
 	} // END get_public_item_schema()
-
-	/**
-	 * Get the query params for getting the cart.
-	 *
-	 * @access public
-	 *
-	 * @since 2.1.0 Introduced.
-	 * @since 4.0.0 Added `config` parameters.
-	 *
-	 * @return array $params The query params.
-	 */
-	public function get_collection_params() {
-		$params = array(
-			'cart_key' => array(
-				'description'       => __( 'Unique identifier for the cart.', 'cart-rest-api-for-woocommerce' ),
-				'type'              => 'string',
-				'required'          => false,
-				'sanitize_callback' => 'sanitize_key',
-				'validate_callback' => 'rest_validate_request_arg',
-			),
-			'config'   => array(
-				'description' => __( 'Configure the cart response for each sub-parameter.', 'cart-rest-api-for-woocommerce' ),
-				'type'        => 'object',
-				'required'    => false,
-				'items'       => array(
-					'type'       => 'object',
-					'properties' => array(
-						'fields' => array(
-							'description'       => __( 'Specify the type of cart response before the data is fetched.', 'cart-rest-api-for-woocommerce' ),
-							'type'              => 'string',
-							'required'          => false,
-							'enum'              => array(
-								'digital',
-								'digital_fees',
-								'shipping',
-								'shipping_fees',
-								'removed_items',
-								'cross_sells',
-							),
-							'sanitize_callback' => 'sanitize_key',
-							'validate_callback' => 'rest_validate_request_arg',
-						),
-						'prices' => array(
-							'description'       => __( 'Return the price values in the format you prefer.', 'cart-rest-api-for-woocommerce' ),
-							'type'              => 'string',
-							'required'          => false,
-							'enum'              => array( 'preformatted' ),
-							'sanitize_callback' => 'sanitize_key',
-							'validate_callback' => 'rest_validate_request_arg',
-						),
-					),
-				),
-			),
-			'fields'   => array(
-				'description'       => __( 'Specify each parent field you want to request separated by (,) in the cart response before the data is fetched.', 'cart-rest-api-for-woocommerce' ),
-				'type'              => 'string',
-				'required'          => false,
-				'sanitize_callback' => 'sanitize_key',
-				'validate_callback' => 'rest_validate_request_arg',
-			),
-			'thumb'    => array(
-				'description' => __( 'True if you want to return the URL of the featured product image for each item in the cart.', 'cart-rest-api-for-woocommerce' ),
-				'default'     => true,
-				'type'        => 'boolean',
-				'required'    => false,
-			),
-			'default'  => array(
-				'description' => __( 'Return the default cart data if set to true.', 'cart-rest-api-for-woocommerce' ),
-				'default'     => false,
-				'type'        => 'boolean',
-				'required'    => false,
-			),
-		);
-
-		/**
-		 * Extend the query parameters for the cart.
-		 *
-		 * This filter allows you to extend the query parameters without removing any default parameters.
-		 *
-		 * @since 3.1.0 Introduced.
-		 */
-		$params += apply_filters( 'cocart_cart_query_parameters', array() );
-
-		return $params;
-	} // END get_collection_params()
 
 } // END class
